chore: address feedback from review
